+++ packages/runtime/plugin-runtime/src/core/plugin/base.ts
@@ -1,52 +1,61 @@
 import {
   createManager,
-  createPipeline,
-  createAsyncPipeline,
+  createWaterfall,
+  createAsyncInterruptWorkflow,
+  createSyncParallelWorkflow,
   PluginOptions,
   Setup,
+  createContext,
 } from '@modern-js/plugin';
 
 import { RuntimeContext, TRuntimeContext } from '../context/runtime';
+import type { RuntimeConfig } from './index';
+
+export const RuntimeConfigContext = createContext<RuntimeConfig>({});
+
+export const useRuntimeConfigContext = () => RuntimeConfigContext.use().value;
 
 // eslint-disable-next-line @typescript-eslint/no-empty-interface
 export interface AppProps {}
 
-const hoc = createPipeline<
-  {
-    App: React.ComponentType<any>;
-    config: Record<string, any>;
-  },
-  React.ComponentType<any>
->();
+const wrapRoot = createWaterfall<React.ComponentType<any>>();
 
-const init = createAsyncPipeline<
-  {
-    context: RuntimeContext;
-  },
-  unknown
->();
+const beforeRender = createAsyncInterruptWorkflow<RuntimeContext, void>();
 
 /**
  * To add runtime info to runtime context
  */
-const pickContext = createPipeline<
-  { context: RuntimeContext; pickedContext: TRuntimeContext },
-  TRuntimeContext
+const pickContext = createWaterfall<TRuntimeContext>();
+
+const modifyRuntimeConfig = createSyncParallelWorkflow<
+  void,
+  Record<string, any>
 >();
 
 const runtimeHooks = {
-  hoc,
-  init,
+  beforeRender,
+  wrapRoot,
   pickContext,
+  modifyRuntimeConfig,
+};
+
+const runtimePluginAPI = {
+  useRuntimeConfigContext,
 };
 
 /** All hooks of runtime plugin. */
 export type RuntimeHooks = typeof runtimeHooks;
 
-/** Plugin options of a runtime plugin. */
-export type Plugin = PluginOptions<RuntimeHooks, Setup<RuntimeHooks>>;
+export type RuntimePluginAPI = typeof runtimePluginAPI;
 
-export const createRuntime = () => createManager(runtimeHooks);
+/** Plugin options of a runtime plugin. */
+export type Plugin = PluginOptions<
+  RuntimeHooks,
+  Setup<RuntimeHooks, RuntimePluginAPI>
+>;
+
+export const createRuntime = () =>
+  createManager(runtimeHooks, runtimePluginAPI);
 
 export const runtime = createRuntime();