+++ packages/cli/core/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @modern-js/core
 
+## 2.57.0
+
+### Patch Changes
+
+- Updated dependencies [2515b00]
+- Updated dependencies [dc736ef]
+  - @modern-js/utils@2.57.0
+  - @modern-js/plugin@2.57.0
+  - @modern-js/node-bundle-require@2.57.0
+
 ## 2.56.2
 
 ### Patch Changes