+++ packages/cli/plugin-changeset/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @modern-js/plugin-changeset
 
+## 2.57.0
+
+### Patch Changes
+
+- Updated dependencies [2515b00]
+  - @modern-js/utils@2.57.0
+  - @modern-js/plugin-i18n@2.57.0
+
 ## 2.56.2
 
 ### Patch Changes