+++ packages/cli/plugin-lint/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @modern-js/plugin-lint
 
+## 2.57.0
+
+### Patch Changes
+
+- Updated dependencies [2515b00]
+  - @modern-js/utils@2.57.0
+  - @modern-js/tsconfig@2.57.0
+
 ## 2.56.2
 
 ### Patch Changes