+++ packages/cli/plugin-proxy/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @modern-js/plugin-proxy
 
+## 2.57.0
+
+### Patch Changes
+
+- Updated dependencies [2515b00]
+  - @modern-js/utils@2.57.0
+
 ## 2.56.2
 
 ### Patch Changes