+++ packages/generator/generator-common/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @modern-js/generator-common
 
+## 3.4.12
+
+### Patch Changes
+
+- @modern-js/plugin-i18n@2.57.0
+
 ## 3.4.11
 
 ### Patch Changes