+++ packages/generator/generators/router-v5-generator/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @modern-js/router-v5-generator
 
+## 3.4.12
+
+### Patch Changes
+
+- Updated dependencies [2515b00]
+  - @modern-js/utils@2.57.0
+
 ## 3.4.11
 
 ### Patch Changes